updated scripts

+++ _clean.sh
@@ -1,4 +1,8 @@
-#! /bin/bash
+#!/bin/bash
+if [[ $EUID -ne 0 ]]; then
+   echo "must be root" 
+   exit 1
+fi
 
 echo 'terminating containers'
 (cd localstack-compose && docker-compose kill && docker-compose rm -f)
@@ -7,3 +11,4 @@ echo 'removing files'
 rm -rfv ./localstack-compose/volume/tmp
 rm -rfv ./src/.dist
 find . -name '*.tfstate*' -print -delete
+find . -name '*.terraform.lock.hcl' -print -delete
