+++ _destroy.sh
@@ -1,4 +1,8 @@
-#! /bin/bash
+#!/bin/bash
+if [[ $EUID -ne 0 ]]; then
+   echo "must be root" 
+   exit 1
+fi
 
 echo 'destroying resources'
 (cd src/.deploy/consumer-lambda && terraform destroy -auto-approve)
